# src/zenml/zen_server/openapi3_0.yaml
openapi: 3.0.2
info:
  title: ZenML
  version: 0.xx.0
  description: >-
    The Zen API allows access to ZenML programmatically. These docs DO NOT represent the status quo but instead are the Design that is being aimed at.
...
paths:
...
  /stacks:
    get:
      summary: List stacks
      tags:
        - stack
      description: |-
        Get a list of stacks.
      operationId: get_stacks      
      parameters:        
        - in: query
          name: project_name
          required: false
          schema:
            title: Project Name
            type: string
          description: >-
            Stacks are  defined within the project scope. It is
            expected that this endpoint will be used at project scope most of 
            the time.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stacks
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /stacks/{stack_id}:
    get:
      summary: Get a specific stack
      tags:
        - stack
      description: |-
        Get a specific stack using its unique id.
      operationId: get_stack_by_id
      parameters:
        - $ref: '#/components/parameters/stackId'
      responses:
        '200':
          description: Returns the requested stack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
    put:
      summary: Update a specific stack
      tags:
        - stack
      description: |-
        Updates the attribute on a stack.
      operationId: update_stack_by_id
      parameters:
        - $ref: '#/components/parameters/stackId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Stack'
        required: true
      responses:
        '200':
          description: Returns the updated stack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    delete:
      summary: Delete a specific stack
      tags:
        - stack
      description: |-
        Deletes a stack.
      operationId: delete_stack_by_id
      parameters:
        - $ref: '#/components/parameters/stackId'
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /stack/{stack_id}/components:
    get:
      summary: Get components that are part of the stack
      tags:
        - stack
      description: |-
        Returns a list of stack components within the stack.
      operationId: get_components_in_stack
      parameters:
        - $ref: '#/components/parameters/stackId'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components in Stack
                type: array
                items:
                  $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  #-------------------------------
  # Stack Components
  #-------------------------------
...
  /components/{component_type}:
    get:
      summary: Get a list of all stack component types
      tags:
        - component
      description: |-
        Returns a list of stack components 
      operationId: get_component_type_schema    
      parameters:
        - $ref: '#/components/parameters/componentType'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components
                type: array
                items:
                  $ref: '#/components/schemas/ComponentType'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []    
...
  /components/{component_id}:
    get:
      summary: Get a specific stack component
      tags:
        - component
      description: |-
        Get a specific stack component using its unique id.
      operationId: get_component_by_id
      parameters:
        - $ref: '#/components/parameters/componentId'
      responses:
        '200':
          description: Returns the requested component.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
    put:
      summary: Update a specific component
      tags:
        - component
      description: |-
        Updates the attribute on a component.
      operationId: update_component_by_id
      parameters:
        - $ref: '#/components/parameters/componentId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Component'
        required: true
      responses:
        '200':
          description: Returns the updated component.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    delete:
      summary: Delete a specific component
      tags:
        - component
      description: |-
        Deletes a project.
      operationId: delete_component_by_id
      parameters:
        - $ref: '#/components/parameters/componentId'
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /projects/{project_name}/stacks:
    get:
      summary: Get stacks that are part of the project
      tags:
        - project
      description: |-
        Returns all stacks within the project.
      operationId: get_stacks_in_project
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stacks in Project
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create a stack
      tags:
        - project
      description: |-
        Creates a stack based on the requestBody.
      operationId: create_stack_in_project
      parameters:
        - $ref: '#/components/parameters/projectName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Stack'
        required: true
      responses:
        '200':
          description: Returns the created Stack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /projects/{project_name}/components:
    get:
      summary: Get components of all types that are part of the project
      tags:
        - project
      description: |-
        Returns all stack components within the project.
      operationId: get_components_in_project
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components in Project
                type: array
                items:
                  $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /projects/{project_name}/components/{component_type}:
    get:
      summary: Get components of type that are part of the project
      tags:
        - project
      description: |-
        Returns all stack components within the project.
      operationId: get_components_of_type_in_project
      parameters:
        - $ref: '#/components/parameters/componentType'
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components in Project
                type: array
                items:
                  $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create a component of type
      tags:
        - project
      description: |-
        Creates a component of type based on the requestBody.
      operationId: create_component_of_type_in_project
      parameters:
        - $ref: '#/components/parameters/componentType'
        - $ref: '#/components/parameters/projectName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Component'
        required: true
      responses:
        '200':
          description: Returns the created Component.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
components:
...
  schemas:
...
    Stack:
      title: Stack
      required:
        - id
        - name
        - components
      type: object
      properties:
        id:
          title: id
          type: string
          readOnly: True
          example: 9lg2jk3
        name:
          title: Name
          type: string
          example: prod_stack
        components:
          title: Components
          type: array
          items:
            $ref: '#/components/schemas/Component'
        creator:
          title: Name
          type: string
          example: Hamza Tahir
        created_at:
          title: Name
          type: string
          example: "2022-08-12T07:12:45.931Z"
      description: >-
        Description of a stack which is a combination of multiple stack
        components.
    Component:
      title: Component
      required:
        - type
        - flavor
        - name
        - id
        - config
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ComponentType'
        flavor:
          title: Flavor
          type: string
        name:
          title: Name
          type: string
        id:
          title: id
          type: string
          format: uuid
          readOnly: True
        config:
          type: object
          oneOf:
            - $ref: '#/components/schemas/SlackAlerterFlavor'
            - $ref: '#/components/schemas/SeldonDeployerFlavor'
          discriminator:
            propertyName: flavor
      description: Serializable Configuration of a StackComponent.
      example:        
        id: asdf2-asdf1-ad143
        type: alerter
        flavor: slack
        name: slack_alerter
        config: {slack_token: INFU1328UJ, default_slack_channel_id: 184195}
    ComponentType:
      title: ComponentType
      enum:
        - alerter
        - annotator
        - artifact_store
        - container_registry
        - data_validator
        - experiment_tracker
        - feature_store
        - metadata_store
        - model_deployer
        - orchestrator
        - secrets_manager
        - step_operator
      type: string
      description: All possible types a `StackComponent` can have.
...
    SlackAlerterFlavor:
      title: SlackAlerterFlavor
      required:
        - slack_token
      type: object
      properties:
        slack_token:
          title: configuration
          type: string
        default_slack_channel_id:
          title: configuration
          type: string
      description: Serializable Configuration for the Slack Alerter with default values prefilled.
...
    SeldonDeployerFlavor:
      title: SeldonDeployerFlavor
      required:
        - base_url
      type: object
      properties:
        kubernetes_context:
          title: configuration
          type: string
        kubernetes_namespace:
          title: configuration
          type: string
        base_url:
          title: base_url
          type: string
      description: Serializable Configuration for the Seldon Deployer with default values prefilled.
...
    ErrorModel:
      title: ErrorModel
      type: object
      properties:
        detail:
          title: Detail
      description: Base class for error responses.
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
  #-------------------------------
  # Reusable operation parameters
  #-------------------------------
  parameters:
...
    projectName:
      name: project_name
      in: path
      required: true
      schema:
        title: Project Name
        type: string
...
    stackId:
      name: stack_id
      in: path
      required: true
      schema:
        title: Stack Id
        type: string
    componentId:
      name: component_id
      in: path
      required: true
      schema:
        title: Component Id
        type: string
    componentType:
      name: component_type
      in: path
      required: true
      schema:
        title: Component Id
        type: string
        enum:
          - alerter
          - annotator
          - artifact_store
          - container_registry
          - data_validator
          - experiment_tracker
          - feature_store
          - metadata_store
          - model_deployer
          - orchestrator
          - secrets_manager
          - step_operator
  #-------------------------------
  # Reusable responses
  #-------------------------------
  responses:
    '200':
      description: Successful Response
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'        
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '422':
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic